+++ tests/Unit/ValidConfigurationTest.php
@@ -3,8 +3,8 @@
 namespace LaravelFlatfilesTest\Unit;
 
 use LaravelFlatfiles\FlatfileExport;
-use LaravelFlatfilesTest\TestCase;
 use LaravelFlatfiles\FlatfileFields;
+use LaravelFlatfilesTest\TestCase;
 
 class ValidConfigurationTest extends TestCase implements FlatfileFields
 {
@@ -14,17 +14,21 @@ class ValidConfigurationTest extends TestCase implements FlatfileFields
     public function fields()
     {
         return [
-            'field' => 'Label',
-            'relation.field' => [
-                'label' => 'Custom label',
+            'field'                                        => 'Label',
+            'relation.field'                               => [
+                'label'    => 'Custom label',
                 'callback' => function ($field) {
                     return $field;
                 },
             ],
             [
                 'column' => 'special.field',
-                'label' => 'Special Label',
+                'label'  => 'Special Label',
             ],
+            'field_where_label_equals_field',
+            'field_where_label_equals_field_with_callback' => function () {
+                return 'value';
+            },
         ];
     }
 
@@ -41,18 +45,29 @@ class ValidConfigurationTest extends TestCase implements FlatfileFields
         $this->assertEquals([
             [
                 'column' => 'field',
-                'label' => 'Label',
+                'label'  => 'Label',
             ],
             [
-                'column' => 'relation.field',
-                'label' => 'Custom label',
+                'column'   => 'relation.field',
+                'label'    => 'Custom label',
                 'callback' => function ($field) {
                     return $field;
                 },
             ],
             [
                 'column' => 'special.field',
-                'label' => 'Special Label',
+                'label'  => 'Special Label',
+            ],
+            [
+                'column' => 'field_where_label_equals_field',
+                'label'  => 'field_where_label_equals_field',
+            ],
+            [
+                'column'   => 'field_where_label_equals_field_with_callback',
+                'label'    => 'field_where_label_equals_field_with_callback',
+                'callback' => function () {
+                    return 'value';
+                },
             ],
         ], $this->flatfile->configuration()->fields()->values()->toArray());
     }
@@ -60,12 +75,24 @@ class ValidConfigurationTest extends TestCase implements FlatfileFields
     /** @test */
     public function it_knows_all_columns()
     {
-        $this->assertEquals(['field', 'relation.field', 'special.field'], $this->flatfile->configuration()->columns());
+        $this->assertEquals([
+            'field',
+            'relation.field',
+            'special.field',
+            'field_where_label_equals_field',
+            'field_where_label_equals_field_with_callback',
+        ], $this->flatfile->configuration()->columns());
     }
 
     /** @test */
     public function it_knows_all_labels()
     {
-        $this->assertEquals(['Label', 'Custom label', 'Special Label'], $this->flatfile->configuration()->fieldLabels());
+        $this->assertEquals([
+            'Label',
+            'Custom label',
+            'Special Label',
+            'field_where_label_equals_field',
+            'field_where_label_equals_field_with_callback',
+        ], $this->flatfile->configuration()->fieldLabels());
     }
 }